fix few dependency cycles
