+++ packages/grid/x-data-grid/src/models/gridFilterOperator.ts
@@ -1,12 +1,21 @@
 import * as React from 'react';
+import { AutocompleteProps } from '@mui/material/Autocomplete';
 import { GridFilterInputValueProps } from '../components/panel/filterPanel/GridFilterInputValueProps';
 import { GridFilterInputMultipleValueProps } from '../components/panel/filterPanel/GridFilterInputMultipleValue';
-import { GridFilterInputMultipleSingleSelectProps } from '../components/panel/filterPanel/GridFilterInputMultipleSingleSelect';
 import { GridFilterItem } from './gridFilterItem';
 import { GridCellParams } from './params/gridCellParams';
 import type { GridStateColDef } from './colDef/gridColDef';
 import type { GridApiCommunity } from './api/gridApiCommunity';
 import type { GridApiCommon } from './api/gridApiCommon';
+
+export type GridFilterInputMultipleSingleSelectProps = {
+  item: GridFilterItem;
+  applyValue: (value: GridFilterItem) => void;
+  // Is any because if typed as GridApiRef a dep cycle occurs. Same happens if ApiContext is used.
+  apiRef: React.MutableRefObject<GridApiCommon>;
+  focusElementRef?: React.Ref<any>;
+  type?: 'singleSelect';
+} & Omit<AutocompleteProps<any[], true, false, true>, 'options' | 'renderInput'>;
 
 /**
  * Filter operator definition interface.
